Troubleshooting section
Updated the README with some error messages that came up during the course of trying this out for the first time on an existing API Platform implementation. Most problems were due to using an incorrect entrypoint to start with and then having authentication on the entrypoint and context routes. Hopefully this will save someone else the time in the future.